requirements.txt
attrs==19.3.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
google-api-core==1.17.0
google-api-python-client==1.9.1
google-auth==1.16.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.9
importlib-metadata==1.6.1
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.3.0
mypy==0.780
mypy-extensions==0.4.3
oauth2client==3.0.0
oauthlib==3.1.0
packaging==20.4
phonenumbers==8.12.5
pluggy==0.13.1
protobuf==3.12.2
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.3
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.15.0
twilio==6.41.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.3
zipp==3.1.0
